mutations file code reduced
